Confirmation Boxes
- A confirmation box has been added to DatabaseView.Delete()
- A confirmation box has been added to StudentForm.Cancel()
- Mailing address hidden unless "seperatemailingaddressbox" checked
"yes"
- Added "newAppButton" to DatabaseView
- Rescaled Hover animations

+++ DatabaseView.xaml.cs
@@ -1,10 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Windows;
+﻿using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Data;
 using System.Windows.Input;
-using System.Windows.Media;
 
 namespace Sappi
 {
@@ -13,10 +10,6 @@ namespace Sappi
     /// </summary>
     public partial class DatabaseView : UserControl
     {
-        //public static readonly DependencyProperty CurrentColumnProperty = DependencyProperty.Register(
-        //"CurrentColumn", typeof(string), typeof(DatabaseView), new PropertyMetadata("Item 1"));
-        ComboBox statusBox = new ComboBox();
-
         public string CurrentColumn { get; set; }
         
         public DatabaseView()
@@ -24,19 +17,13 @@ namespace Sappi
             InitializeComponent();
             //root.DataContext = this;
             dg.MouseRightButtonUp += Context_MouseRightButtonUp;
-            statusBox.SelectionChanged += SetStatus;
             CurrentColumn = " ";
-            //NameScope.SetNameScope(ContextMenu, NameScope.GetNameScope(this));
-        }
-        
+        }        
 
         private void DatabaseView_Loaded(object sender, RoutedEventArgs e)
         {
-            //student_contentArea.Content = new StudentView();
             //set the databinding for DataGrid
             dg.ItemsSource = App.db.items;
-            statusBox.ItemsSource = App.formData.status.Keys;
-            statusBox.IsEnabled = false;
             PopulateDatagrid(dg);
         }
 
@@ -65,7 +52,6 @@ namespace Sappi
         }
         private void RowDeleteCommand_Executed(object sender, ExecutedRoutedEventArgs e)
         {
-            //are you sure you want to delete?
             Delete();
         }
         private void RowDeleteCommand_CanExecute(object sender, CanExecuteRoutedEventArgs e)
@@ -144,16 +130,6 @@ namespace Sappi
                 Binding = new Binding("groupBoxes[9]")
                 { Converter = new IntToKeyConverter() }
             });
-            //for (int i = 0; i < App.formData.groups.Count; ++i)
-            //{
-            //    DataGridColumn c;
-            //    c.Header = 
-            //    dg.Columns.Add(c);
-            //    for (int j = 0; j < App.db.items.Count; ++j)
-            //    {
-
-            //    }
-            //}
         }
 
         private void Edit()
@@ -163,37 +139,35 @@ namespace Sappi
 
         private void Delete()
         {
-            for (int i = 0; i < dg.SelectedItems.Count; ++i)
+            MessageBoxResult messageBoxResult = MessageBox.Show("Are you sure?", 
+                                                "Delete Confirmation", MessageBoxButton.YesNo);
+            if (messageBoxResult == MessageBoxResult.Yes)
             {
-                App.db.items.Remove((StudentData) dg.SelectedItems[i]);
-            }
-            App.db.items.Remove((StudentData)dg.SelectedItem);
-            dg.Items.Refresh();
-        }
-
-        private void SetStatus(object sender, SelectionChangedEventArgs e)
-        {
-            //(dg.SelectedItem as StudentData).status = e.OriginalSource;
+                for (int i = 0; i < dg.SelectedItems.Count; ++i)
+                {
+                    App.db.items.Remove((StudentData)dg.SelectedItems[i]);
+                }
+                App.db.items.Remove((StudentData)dg.SelectedItem);
+                dg.Items.Refresh();
+            }            
         }
 
         private void Search(object sender, TextChangedEventArgs e)
         {
             if (searchBar.Text == "" || searchBar.Text == "(search)")
                 return;
-
-            //this is used so the full name doesn't need to be entered
-            //in the search bar before finding results
+            
             int inputLength = searchBar.Text.Length;
 
-            //loop through rows (students)
+            //loop through rows
             for (int i = 0; i < dg.Items.Count; ++i)
             {
-                //get the Name cell in each row
+                //get the "Name" cell in each row
                 DataGridCell cell = DataGridHelper.GetCell(dg, i, 0);
 
                 //get text content
                 var cellContent = (TextBlock)cell.Content;
-
+                //check for out of range exception
                 if (cellContent.Text.Length < inputLength)
                     break;
 
@@ -204,9 +178,10 @@ namespace Sappi
                     dg.SelectedItem = dg.Items[i];
             }
         }
-        //empty the text box of default text when clicked
+        
         private void Textbox_GotFocus(object sender, RoutedEventArgs e)
         {
+            //empty the text box of default text when clicked
             TextBox tb = (TextBox)sender;
             tb.Text = string.Empty;
             tb.GotFocus -= Textbox_GotFocus;
@@ -215,77 +190,32 @@ namespace Sappi
 
         private void EditStudentData(object sender, MouseButtonEventArgs e)
         {
-            //MainWindow.Main.ContentArea.Content = new StudentForm(dg.SelectedItem as StudentData);
-            if(e.ChangedButton == MouseButton.Left)
-                Edit();
-        }
-        private void GoBack(object sender, RoutedEventArgs e)
-        {
-            //can't set equal to App.previous because if entered a StudentForm then
-            //at this point App.previous is equal to DatabaseView, going no where
-            MainWindow.Main.ContentArea.Content = new StartUp();
+            Edit();
         }
 
         private void Datagrid_KeyUp(object sender, KeyEventArgs e)
         {
             if(e.Key == Key.Delete)
-            {
                 Delete();
-            }
         }
-
-        //private void editButton_Click(object sender, RoutedEventArgs e)
-        //{
-        //    if (studentInfo_contentArea.Content == null)
-        //    {
-        //        studentInfo_contentArea.Content = new StudentForm();
-        //        editButton.Content = "Done";
-        //        //disable datagrid so that studentform doesn't visually overlap
-
-        //    }
-        //    else
-        //    {
-        //        studentInfo_contentArea.Content = null;
-        //        editButton.Content = "Edit";
-        //    }
-        //}
 
         private void Context_MouseRightButtonUp(object sender, MouseEventArgs e)
         {
             DependencyObject dep = (DependencyObject)e.OriginalSource;
-            if (dep != null)
+            if (dep is TextBlock)
             {
-                //if (dep is DataGridCell)
-                //{
-                //    DataGridCell cell = dep as DataGridCell;
-                //    cell.Focus();
-                //    Console.WriteLine(cell.Name);
-                //}
                 DataGridCell cell = (DataGridCell)(dep as TextBlock).Parent;
                 if(cell != null)
                     CurrentColumn = cell.Column.Header.ToString();
             }
-
-            //MouseButtonEventArgs arg = new MouseButtonEventArgs(Mouse.PrimaryDevice, 0, MouseButton.Left);
-            //arg.RoutedEvent = MouseLeftButtonDownEvent;
-            //dg.RaiseEvent(arg);
-
-            //HitTestResult htr = dg.InputHitTest(e.GetPosition(dg));// .HitTest(dg, e.GetPosition(dg));
-            //DataGridRow row = htr.VisualHit.DependencyObjectType<DataGridRow>();
-            //if (e.OriginalSource != null)
-            //{
-            //    StudentData s = ((e.OriginalSource as TextBlock).DataContext) as StudentData;
-            //}
-            StudentData ss = (StudentData)dg.CurrentItem;
-
-
-            var row = (sender as DataGrid).SelectedItem;
-            //dg.SelectedItem = row;
-            //Console.WriteLine(ss.name);
         }
 
         private void contextMenu_Opened(object sender, RoutedEventArgs e)
         {
+            // |item_lvl1|
+            // |item_lvl1| -> |item_lvl2|
+            //                |item_lvl2|
+
             ContextMenu menu = (ContextMenu)e.OriginalSource;
             
             MenuItem item1_lvl1 = (MenuItem)menu.Items[0];
@@ -302,8 +232,7 @@ namespace Sappi
 
                     item1_lvl1.Items.Add(item_lvl2);
                 }
-            }
-            
+            }            
         }
         private void contextMenu_Closed(object sender, RoutedEventArgs e)
         {
@@ -312,5 +241,20 @@ namespace Sappi
 
             item1_lvl1.Items.Clear();
         }
+
+        private void backButton_Click(object sender, RoutedEventArgs e)
+        {
+            //     Can't set equal to App.previous because if user entered a StudentForm,
+            // at this point App.previous == current (DatabaseView). Perhaps in the
+            // future, save a history of previous states so that if App.previous[0] ==
+            // current, set content = App.previous[1]
+            MainWindow.Main.ContentArea.Content = new StartUp();
+        }
+
+        private void newAppButton_Click(object sender, RoutedEventArgs e)
+        {
+            MainWindow.ChangeState(App.Page.StudentForm);
+        }
     }
+
 }
